backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# WebSockets
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# AI / ML
openai-whisper==20231117
torch>=2.0.0
numpy<2

# Translation
deep-translator==1.11.4

# Text-to-Speech
edge-tts==6.1.9

# Language Detection
langdetect==1.0.9

# Audio Processing
pydub==0.25.1
soundfile==0.12.1

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2
